+++ sdk/src/wallet/account/operations/balance.rs
@@ -28,10 +28,10 @@ impl Account {
 
         let account_addresses = self.addresses().await?;
 
-        let network_id = self.client.get_network_id().await?;
-        let rent_structure = self.client.get_rent_structure().await?;
+        let network_id = self.client().get_network_id().await?;
+        let rent_structure = self.client().get_rent_structure().await?;
 
-        let local_time = self.client.get_time_checked().await?;
+        let local_time = self.client().get_time_checked().await?;
 
         let mut total_rent_amount = 0;
         let mut total_native_tokens = NativeTokensBuilder::new();