+++ sdk/src/wallet/account/operations/syncing/transactions.rs
@@ -37,7 +37,7 @@ impl Account {
             return Ok(confirmed_unknown_output);
         }
 
-        let network_id = self.client.get_network_id().await?;
+        let network_id = self.client().get_network_id().await?;
 
         let mut updated_transactions = Vec::new();
         let mut spent_output_ids = Vec::new();
@@ -98,7 +98,7 @@ impl Account {
             }
 
             if let Some(block_id) = transaction.block_id {
-                match self.client.get_block_metadata(&block_id).await {
+                match self.client().get_block_metadata(&block_id).await {
                     Ok(metadata) => {
                         if let Some(inclusion_state) = metadata.ledger_inclusion_state {
                             match inclusion_state {
@@ -120,7 +120,7 @@ impl Account {
                                     // try to get the included block, because maybe only this attachment is
                                     // conflicting because it got confirmed in another block
                                     if let Ok(included_block) =
-                                        self.client.get_included_block(&transaction.payload.id()).await
+                                        self.client().get_included_block(&transaction.payload.id()).await
                                     {
                                         confirmed_unknown_output = true;
                                         updated_transaction_and_outputs(