+++ sdk/src/wallet/wallet/operations/address_generation.rs
@@ -125,16 +125,7 @@ impl Wallet {
                     .address
                     .hrp
             }
-            None => {
-                self.client_options
-                    .read()
-                    .await
-                    .clone()
-                    .finish()
-                    .await?
-                    .get_bech32_hrp()
-                    .await?
-            }
+            None => self.client().get_bech32_hrp().await?,
         })
     }
 }